enemysprite added

+++ Sprite.java
@@ -18,9 +18,9 @@ public abstract class Sprite implements Comparable<Sprite>
     
     public int getY() {return yPos;}
 
-    public void setX(double newX) {xPos = (int)Math.round(newX);}
+    public void setX(float newX) {xPos = (int)Math.round(newX);}
     
-    public void setY(double newY) {yPos = (int)Math.round(newY);}
+    public void setY(float newY) {yPos = (int)Math.round(newY);}
     
     public int getWidth() {return width;}
